debug JQuery importing and google map api errors

+++ config/application.rb
@@ -15,9 +15,11 @@ module Sharedesk
     # setting env variable for development mode only
     config.before_configuration do
       env_file = File.join(Rails.root, 'config', 'local_env.yml')
-      YAML.load(File.open(env_file)).each do |key, value|
-        ENV[key.to_s] = value
-      end if File.exist?(env_file)
+      if File.exist?(env_file)
+        YAML.load(File.open(env_file)).each do |key, value|
+          ENV[key.to_s] = value.to_s
+        end
+      end
     end
